@@ -13,7 +13,9 @@
  * License along with MoFEM. If not, see <http://www.gnu.org/licenses/>. */
 
 #ifndef NDEBUG
-#define BOOST_DISABLE_ASSERTS
+#define BOOST_DISABLE_ASSERTS ///< If not set makes BOOST freezing, check what
+                              ///< happen fi boost recompiled with cxx17
+                              ///< standard
 #define BOOST_MULTI_INDEX_ENABLE_INVARIANT_CHECKING
 #define BOOST_MULTI_INDEX_ENABLE_SAFE_MODE
 #endif
