@@ -288,7 +288,7 @@ PetscErrorCode ShapeMBTETQ_inverse(double *N, double *diffN,
 // complex part
 void ShapeDiffMBTETinvJ_complex(double *diffN, __CLPK_doublecomplex *invJac,
                                 __CLPK_doublecomplex *diffNinvJac,
-                                const CBLAS_TRANSPOSE Trans);
+                                CBLAS_TRANSPOSE Trans);
 PetscErrorCode ShapeFaceNormalMBTRI_complex(double *diffN,
                                             __CLPK_doublecomplex *xcoords,
                                             __CLPK_doublecomplex *xnormal);
